🔨 Refactor: 추상화 제거, 사용자 역할 오타 수정
- 장바구니 Service 추상화 제거
- 장바구니 Repository 통합을 위해 제거
- 유저 역할 오타 수정
