--- src/main/java/com/example/cloudfour/peopleofdelivery/domain/cart/repository/CartRepository.java
@@ -1,4 +1,0 @@
-package com.example.cloudfour.peopleofdelivery.domain.cart.repository;
-
-public interface CartRepository {
-}
