+++ src/main/java/com/example/cloudfour/peopleofdelivery/domain/cart/service/command/CartCommandService.java
@@ -3,7 +3,21 @@ package com.example.cloudfour.peopleofdelivery.domain.cart.service.command;
 import com.example.cloudfour.peopleofdelivery.domain.cart.dto.CartRequestDTO;
 import com.example.cloudfour.peopleofdelivery.domain.cart.dto.CartResponseDTO;
 import com.example.cloudfour.peopleofdelivery.domain.user.entity.User;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
 
 import java.util.UUID;
+@Service
+@Slf4j
+@RequiredArgsConstructor
+@Transactional
+public class CartCommandService {
+    public CartResponseDTO.CartCreateResponseDTO createCart(CartRequestDTO.CartCreateRequestDTO CartCreateRequestDTO, User user) {
+        return null;
+    }
 
+    public void deleteCart(UUID cartId, User user) {
+    }
 }
